Renamed files + added detailed comments.
