+++ compute/client_library/recipes/compute_reservations/consume_single_project_reservation.py
