+++ compute/client_library/snippets/compute_reservations/consume_single_project_reservation.py
@@ -12,15 +12,70 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# This is an ingredient file. It is not meant to be run directly. Check the samples/snippets
-# folder for complete code samples that are ready to be used.
-# Disabling flake8 for the ingredients file, as it would fail F821 - undefined name check.
-# flake8: noqa
 
+# This file is automatically generated. Please do not modify it directly.
+# Find the relevant recipe file in the samples/recipes or samples/ingredients
+# directory and apply your changes there.
+
+
+# [START compute_consume_single_project_reservation]
+from __future__ import annotations
+
+import sys
+from typing import Any
+
+from google.api_core.extended_operation import ExtendedOperation
 from google.cloud import compute_v1
 
 
-# <INGREDIENT consume_single_project_reservation>
+def wait_for_extended_operation(
+    operation: ExtendedOperation, verbose_name: str = "operation", timeout: int = 300
+) -> Any:
+    """
+    Waits for the extended (long-running) operation to complete.
+
+    If the operation is successful, it will return its result.
+    If the operation ends with an error, an exception will be raised.
+    If there were any warnings during the execution of the operation
+    they will be printed to sys.stderr.
+
+    Args:
+        operation: a long-running operation you want to wait on.
+        verbose_name: (optional) a more verbose name of the operation,
+            used only during error and warning reporting.
+        timeout: how long (in seconds) to wait for operation to finish.
+            If None, wait indefinitely.
+
+    Returns:
+        Whatever the operation.result() returns.
+
+    Raises:
+        This method will raise the exception received from `operation.exception()`
+        or RuntimeError if there is no exception set, but there is an `error_code`
+        set for the `operation`.
+
+        In case of an operation taking longer than `timeout` seconds to complete,
+        a `concurrent.futures.TimeoutError` will be raised.
+    """
+    result = operation.result(timeout=timeout)
+
+    if operation.error_code:
+        print(
+            f"Error during {verbose_name}: [Code: {operation.error_code}]: {operation.error_message}",
+            file=sys.stderr,
+            flush=True,
+        )
+        print(f"Operation ID: {operation.name}", file=sys.stderr, flush=True)
+        raise operation.exception() or RuntimeError(operation.error_message)
+
+    if operation.warnings:
+        print(f"Warnings during {verbose_name}:\n", file=sys.stderr, flush=True)
+        for warning in operation.warnings:
+            print(f" - {warning.code}: {warning.message}", file=sys.stderr, flush=True)
+
+    return result
+
+
 def consume_specific_single_project_reservation(
     project_id: str,
     zone: str,
@@ -30,7 +85,8 @@ def consume_specific_single_project_reservation(
     min_cpu_platform: str = "Intel Ivy Bridge",
 ) -> compute_v1.Instance:
     """
-    Create a specific single project reservation and an instance that targets the reservation.
+    Creates a specific reservation in a single project and launches a VM
+    that consumes the newly created reservation.
     Args:
         project_id (str): The ID of the Google Cloud project.
         zone (str): The zone to create the reservation.
@@ -113,4 +169,4 @@ def consume_specific_single_project_reservation(
     return vm_client.get(project=project_id, zone=zone, instance=instance_name)
 
 
-# </INGREDIENT>
+# [END compute_consume_single_project_reservation]
